refactor(editor): Extract commandBar code into features
