+++ packages/frontend/editor-ui/src/features/ui/commandBar/commandBar.types.ts
